+++ src/commands/sfpowerkit/source/profile/sync.ts
@@ -2,29 +2,28 @@ import { core, SfdxCommand, flags, FlagsConfig } from "@salesforce/command";
 
 import { SfdxProject } from "@salesforce/core";
 import _ from "lodash";
-import AcnProfileUtils from "../../../profile_utils/profileUtils";
-import { SfPowerKit } from "../../../shared/sfpowerkit";
+import AcnProfileUtils from "../../../../profile_utils/profileUtils";
+import { SfPowerKit } from "../../../../shared/sfpowerkit";
 
 // Initialize Messages with the current plugin directory
 core.Messages.importMessagesDirectory(__dirname);
 
 // Load the specific messages for this file. Messages from @salesforce/command, @salesforce/core,
 // or any library that is using the messages framework can also be loaded this way.
-const messages = core.Messages.loadMessages("sfpowerkit", "profile_reconcile");
+const messages = core.Messages.loadMessages("sfpowerkit", "profile_sync");
 
-export default class Reconcile extends SfdxCommand {
+export default class Sync extends SfdxCommand {
   public static description = messages.getMessage("commandDescription");
 
   public static examples = [
-    `$ sfdx sfpowerkit:profile:reconcile  --folder force-app`,
-    `$ sfdx sfpowerkit:profile:reconcile  --folder force-app,module2,module3 -u sandbox`,
-    `$ sfdx sfpowerkit:profile:reconcile  -u myscratchorg`
+    `$ sfdx sfpowerkit:source:profile:sync -u prod`,
+    `$ sfdx sfpowerkit:source:profile:sync  -f force-app -n "My Profile" -r -u prod`,
+    `$ sfdx sfpowerkit:source:profile:sync  -f "module1, module2, module3" -n "My Profile1, My profile2"  -u prod`
   ];
 
-  //public static args = [{name: 'file'}];
+  //public static args = [{ name: 'file' }];
 
   protected static flagsConfig: FlagsConfig = {
-    // flag with a value (-n, --name=VALUE)
     folder: flags.array({
       char: "f",
       description: messages.getMessage("folderFlagDescription"),
@@ -33,35 +32,31 @@ export default class Reconcile extends SfdxCommand {
     }),
     profilelist: flags.array({
       char: "n",
-      description: messages.getMessage("nameFlagDescription"),
+      description: messages.getMessage("profileListFlagDescription"),
       required: false,
-      map: (n: string) => n.trim()
+      map: (p: string) => p.trim()
     })
   };
 
   // Comment this out if your command does not require an org username
   protected static requiresUsername = true;
 
-  // Comment this out if your command does not support a hub org username
-  //protected static supportsDevhubUsername = true;
-
   // Set this to true if your command requires a project workspace; 'requiresProject' is false by default
-  //protected static requiresProject = false;
+  protected static requiresProject = true;
   public async run(): Promise<any> {
-    // tslint:disable-line:no-any
     SfPowerKit.ux = this.ux;
 
-    let argFolder = this.flags.folder;
-    let argProfileList = this.flags.profilelist;
+    let argFolder: string = this.flags.folder;
+    let argProfileList: string[] = this.flags.profilelist;
 
+    let folders: string[] = [];
     if (_.isNil(argFolder) || argFolder.length === 0) {
-      argFolder = [];
       const dxProject = await SfdxProject.resolve();
       const project = await dxProject.retrieveSfdxProjectJson();
 
       let packages = (project.get("packageDirectories") as any[]) || [];
       packages.forEach(element => {
-        argFolder.push(element.path);
+        folders.push(element.path);
         if (element.default) {
           SfPowerKit.defaultFolder = element.path;
         }
@@ -70,13 +65,10 @@ export default class Reconcile extends SfdxCommand {
       SfPowerKit.defaultFolder = argFolder[0];
     }
 
-    var profileUtils = new AcnProfileUtils(this.org);
-    var reconcileProfiles = await profileUtils.reconcile(
-      argFolder,
-      argProfileList || []
-    );
+    const profileUtils = new AcnProfileUtils(this.org);
 
-    // Return an object to be displayed with --json
-    return { ReconciledProfile: reconcileProfiles };
+    let syncPofles = await profileUtils.sync(folders, argProfileList || []);
+
+    return syncPofles;
   }
 }
